backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
openai==1.51.0
PyMuPDF==1.24.10
Pillow==10.4.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
